+++ lib/predicate/nodes/native.rb
@@ -10,10 +10,18 @@ class Predicate
       self[1]
     end
 
+    # overriden because parent relies on free_variables,
+    # which raises an exception
     def and_split(attr_list)
       # I possibly make references to those attributes, so
       # I can't be P2
       [ self, tautology ]
+    end
+
+    # overriden because parent relies on free_variables,
+    # which raises an exception
+    def attr_split
+      { nil => self }
     end
 
     def to_ruby_code(scope = 't')